+++ shapes/cylinder/cylinder.h
@@ -15,13 +15,13 @@ enum INTERSECTION_TYPE {
 class Cylinder : public Shape {
    public:
     Cylinder(Reflexivity reflexivity, Vector3d base_center, Vector3d top_center,
-             double radius);
+             long double radius);
 
-    Cylinder(Reflexivity reflexivity, Vector3d base_center, double height,
-             Vector3d cylinder_direction, double radius);
+    Cylinder(Reflexivity reflexivity, Vector3d base_center, long double height,
+             Vector3d cylinder_direction, long double radius);
 
     Shape * Copy();
-    double intersect(Vector3d &p_0, Vector3d &dr);
+    long double intersect(Vector3d &p_0, Vector3d &dr);
     Vector3d normal(Vector3d &p_i);
     void rebuild_M();
     void operator*(AccMatrix m);
@@ -32,10 +32,10 @@ class Cylinder : public Shape {
     Vector3d top_center_;
     Vector3d cylinder_direction;
     vector<Vector3d> M;
-    double height;
-    double radius_;
-    bool in_cylinder_surface(Vector3d &p0, Vector3d &dr, double &t);
-    bool in_lid_surface(Vector3d &p0, Vector3d &dr, double &t, Vector3d &lid);
+    long double height;
+    long double radius_;
+    bool in_cylinder_surface(Vector3d &p0, Vector3d &dr, long double &t);
+    bool in_lid_surface(Vector3d &p0, Vector3d &dr, long double &t, Vector3d &lid);
     Vector3d *last_dr;
 
     INTERSECTION_TYPE type = INTERSECTION_TYPE::CYLINDER_SURFACE;
