+++ shapes/sphere/sphere.h
@@ -8,15 +8,15 @@
 
 class Sphere : public Shape {
    public:
-    Sphere(Reflexivity reflexivity, Vector3d center, double radius);
+    Sphere(Reflexivity reflexivity, Vector3d center, long double radius);
     Shape * Copy();
-    double intersect(Vector3d &p_0, Vector3d &dr);
+    long double intersect(Vector3d &p_0, Vector3d &dr);
     Vector3d normal(Vector3d &p_i);
     void operator*(AccMatrix m);
     void operator*(gMatrix m);
 
    private:
     Vector3d center_;
-    double radius_;
+    long double radius_;
 };
 #endif
